Address dongjoon-hyun comments
